merian: gbuffer: Use f16 for grad_z

+++ include/merian-shaders/gbuffer.glsl.h
@@ -16,7 +16,7 @@ struct GBuffer {
     // linear distance from camera to pixel
     float linear_z;
     // dlinear_z / dipos in depth / pixel
-    vec2 grad_z;
+    f16vec2 grad_z;
     // camera velocity in ray direction
     float vel_z;
 };
